Event Mapping

+++ app/src/main/java/org/engrave/packup/ui/event/DailyEventsItem.kt
@@ -3,7 +3,7 @@ package org.engrave.packup.ui.event
 import org.engrave.packup.data.deadline.Deadline
 
 data class DailyEventsItem(
-    val startOfDayInMillis: Int,
+    val startOfDayInMillis: Long,
     val deadlines: List<Deadline>,
     val courses: List<DailyCourseItem>
 )
